[수정] 로그아웃 로직 수정, 파일명 변경
